break down and modularize Modal.tsx into smaller components and hooks
