+++ app/packages/core/src/components/Modal/Modal.tsx
@@ -1,25 +1,19 @@
-import { ErrorBoundary, HelpPanel, JSONPanel } from "@fiftyone/components";
 import { OPERATOR_PROMPT_AREAS, OperatorPromptArea } from "@fiftyone/operators";
+import { SpaceNodeJSON, usePanels, useSpaces } from "@fiftyone/spaces";
+import { Space } from "@fiftyone/spaces/src/components";
 import * as fos from "@fiftyone/state";
-import { Controller } from "@react-spring/core";
-import React, {
-  Fragment,
-  Suspense,
-  useCallback,
-  useEffect,
-  useRef,
-} from "react";
+import React, { useCallback, useMemo, useRef } from "react";
 import ReactDOM from "react-dom";
-import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import { useRecoilValue } from "recoil";
 import styled from "styled-components";
-import Sidebar, { Entries } from "../Sidebar";
-import Group from "./Group";
-import { GroupContextProvider } from "./Group/GroupContextProvider";
-import ModalNavigation from "./ModalNavigation";
-import Sample from "./Sample";
-import { Sample3d } from "./Sample3d";
+import Sidebar from "../Sidebar";
+import { NoOpModalContentPluginActivation } from "./ModalContentPlugin";
 import { TooltipInfo } from "./TooltipInfo";
-import { usePanels } from "./hooks";
+import { useModalSidebarRenderEntry } from "./use-sidebar-render-entry";
+
+NoOpModalContentPluginActivation();
+
+export const MODAL_SPACES_ID = "fo-space-modal";
 
 const ModalWrapper = styled.div`
   position: fixed;
@@ -34,7 +28,7 @@ const ModalWrapper = styled.div`
   background-color: ${({ theme }) => theme.neutral.softBg};
 `;
 
-const Container = styled.div`
+const ModalContainer = styled.div`
   background-color: ${({ theme }) => theme.background.level2};
   border: 1px solid ${({ theme }) => theme.primary.plainBorder};
   position: relative;
@@ -44,257 +38,77 @@ const Container = styled.div`
   box-shadow: 0 20px 25px -20px #000;
 `;
 
-const ContentColumn = styled.div`
-  flex-grow: 1;
-  width: 1px;
+const SpacesContainer = styled.div`
+  width: 100%;
   height: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
+  margin-top: 5px;
 `;
 
-const SampleModal = () => {
-  const lookerRef = useRef<fos.Lookers>();
+const Modal = () => {
   const wrapperRef = useRef<HTMLDivElement>(null);
 
-  const disabled = useRecoilValue(fos.fullyDisabledPaths);
-  const labelPaths = useRecoilValue(fos.labelPaths({ expanded: false }));
-
-  const mode = useRecoilValue(fos.groupStatistics(true));
-  const screen = useRecoilValue(fos.fullscreen)
-    ? { width: "100%", height: "100%" }
-    : { width: "95%", height: "90%", borderRadius: "3px" };
-  const isGroup = useRecoilValue(fos.isGroup);
-  const is3D = useRecoilValue(fos.is3DDataset);
   const clearModal = fos.useClearModal();
-  const { jsonPanel, helpPanel, onNavigate } = usePanels();
-  const tooltip = fos.useTooltip();
-  const [isTooltipLocked, setIsTooltipLocked] = useRecoilState(
-    fos.isTooltipLocked
-  );
-  const setTooltipDetail = useSetRecoilState(fos.tooltipDetail);
 
-  const tooltipEventHandler = useCallback(
-    (e) => {
-      if (e.detail) {
-        setTooltipDetail(e.detail);
-        if (!isTooltipLocked && e.detail?.coordinates) {
-          tooltip.setCoords(e.detail.coordinates);
-        }
-      } else if (!isTooltipLocked) {
-        setTooltipDetail(null);
+  const onClickModalWrapper = useCallback(
+    (e: React.MouseEvent) => {
+      if (e.target === wrapperRef.current) {
+        clearModal();
       }
     },
-    [isTooltipLocked, tooltip]
+    [clearModal]
   );
 
-  useEffect(() => {
-    // reset tooltip state when modal is closed
-    setIsTooltipLocked(false);
+  const renderEntry = useModalSidebarRenderEntry();
 
-    return () => {
-      setTooltipDetail(null);
-    };
-  }, []);
+  const isFullScreen = useRecoilValue(fos.fullscreen);
 
-  /**
-   * a bit hacky, this is using the callback-ref pattern to get looker reference so that event handler can be registered
-   * note: cannot use `useEventHandler()` hook since there's no direct reference to looker in Modal
-   */
-  const lookerRefCallback = useCallback(
-    (looker: fos.Lookers) => {
-      lookerRef.current = looker;
-      looker.addEventListener("tooltip", tooltipEventHandler);
-    },
-    [tooltipEventHandler]
+  const screenParams = useMemo(() => {
+    return isFullScreen
+      ? { width: "100%", height: "100%" }
+      : { width: "95%", height: "90%", borderRadius: "3px" };
+  }, [isFullScreen]);
+
+  const panelsPredicate = useCallback(
+    (panel) => panel.surfaces === "modal" || panel.surfaces === "grid modal",
+    []
   );
 
-  const renderEntry = useCallback(
-    (
-      key: string,
-      group: string,
-      entry: fos.SidebarEntry,
-      controller: Controller,
-      trigger: (
-        event: React.MouseEvent<HTMLDivElement>,
-        key: string,
-        cb: () => void
-      ) => void
-    ) => {
-      switch (entry.kind) {
-        case fos.EntryKind.PATH: {
-          const isTag = entry.path === "tags";
-          const isLabelTag = entry.path === "_label_tags";
-          const isLabel = labelPaths.includes(entry.path);
-          const isOther = disabled.has(entry.path);
-          const isFieldPrimitive =
-            !isLabelTag && !isLabel && !isOther && !(isTag && mode === "group");
+  const allModalPlugins = usePanels(panelsPredicate);
 
-          return {
-            children: (
-              <>
-                {(isLabel ||
-                  isOther ||
-                  isLabelTag ||
-                  (isTag && mode === "group")) && (
-                  <Entries.FilterablePath
-                    entryKey={key}
-                    modal={true}
-                    path={entry.path}
-                    group={group}
-                    onFocus={() => {
-                      controller.set({ zIndex: "1" });
-                    }}
-                    onBlur={() => {
-                      controller.set({ zIndex: "0" });
-                    }}
-                    disabled={isOther}
-                    key={key}
-                    trigger={trigger}
-                  />
-                )}
-                {isFieldPrimitive && (
-                  <Entries.PathValue
-                    entryKey={key}
-                    key={key}
-                    path={entry.path}
-                    trigger={trigger}
-                  />
-                )}
-              </>
-            ),
-            disabled: isTag || isOther,
-          };
-        }
-        case fos.EntryKind.GROUP: {
-          return {
-            children: (
-              <Entries.PathGroup
-                entryKey={key}
-                name={entry.name}
-                modal={true}
-                key={key}
-                trigger={trigger}
-              />
-            ),
-            disabled: false,
-          };
-        }
-        case fos.EntryKind.EMPTY:
-          return {
-            children: (
-              <Entries.Empty
-                useText={() => ({ text: "No fields", loading: false })}
-                key={key}
-              />
-            ),
-            disabled: true,
-          };
-        case fos.EntryKind.INPUT:
-          return {
-            children: <Entries.Filter modal={true} key={key} />,
-            disabled: true,
-          };
-        default:
-          throw new Error("invalid entry");
-      }
-    },
-    [disabled, labelPaths, mode]
-  );
+  const defaultModalSpaces = useMemo(() => {
+    return {
+      id: "root",
+      children: allModalPlugins.map((modalPlugin) => ({
+        id: `${modalPlugin.name}`,
+        type: modalPlugin.name,
+        children: [],
+        ...modalPlugin.panelOptions,
+      })),
+      type: "panel-container",
+      // `SampleModal` is the default modal plugin registered in `ModalContentPlugin.tsx`
+      activeChild: "SampleModal",
+    } as SpaceNodeJSON;
+  }, [allModalPlugins]);
 
-  useEffect(() => {
-    return () => {
-      lookerRef.current &&
-        lookerRef.current.removeEventListener("tooltip", tooltipEventHandler);
-    };
-  }, [tooltipEventHandler]);
-
-  const isNestedDynamicGroup = useRecoilValue(fos.isNestedDynamicGroup);
-  const isOrderedDynamicGroup = useRecoilValue(fos.isOrderedDynamicGroup);
-  const isLooker3DVisible = useRecoilValue(fos.groupMedia3dVisibleSetting);
-  const isCarouselVisible = useRecoilValue(
-    fos.groupMediaIsCarouselVisibleSetting
-  );
-
-  const [dynamicGroupsViewMode, setDynamicGroupsViewMode] = useRecoilState(
-    fos.dynamicGroupsViewMode(true)
-  );
-  const setIsMainLookerVisible = useSetRecoilState(
-    fos.groupMediaIsMainVisibleSetting
-  );
-
-  useEffect(() => {
-    // if it is unordered nested dynamic group and mode is not pagination, set to pagination
-    if (
-      isNestedDynamicGroup &&
-      !isOrderedDynamicGroup &&
-      dynamicGroupsViewMode !== "pagination"
-    ) {
-      setDynamicGroupsViewMode("pagination");
-    }
-
-    // hide 3d looker and carousel if `hasGroupSlices`
-    if (
-      dynamicGroupsViewMode === "video" &&
-      (isLooker3DVisible || isCarouselVisible)
-    ) {
-      setIsMainLookerVisible(true);
-    }
-  }, [
-    dynamicGroupsViewMode,
-    isNestedDynamicGroup,
-    isOrderedDynamicGroup,
-    isLooker3DVisible,
-    isCarouselVisible,
-  ]);
+  const { spaces } = useSpaces(MODAL_SPACES_ID, defaultModalSpaces);
 
   return ReactDOM.createPortal(
-    <Fragment>
-      <ModalWrapper
-        ref={wrapperRef}
-        onClick={(event) => event.target === wrapperRef.current && clearModal()}
+    <ModalWrapper ref={wrapperRef} onClick={onClickModalWrapper}>
+      <ModalContainer
+        style={{ ...screenParams, zIndex: 10001 }}
+        data-cy="modal"
       >
-        <Container style={{ ...screen, zIndex: 10001 }} data-cy="modal">
-          <OperatorPromptArea area={OPERATOR_PROMPT_AREAS.DRAWER_LEFT} />
-          <TooltipInfo />
-          <ContentColumn>
-            <ModalNavigation onNavigate={onNavigate} />
-            <ErrorBoundary onReset={() => {}}>
-              <Suspense>
-                {isGroup ? (
-                  <GroupContextProvider lookerRefCallback={lookerRefCallback}>
-                    <Group />
-                  </GroupContextProvider>
-                ) : is3D ? (
-                  <Sample3d />
-                ) : (
-                  <Sample lookerRefCallback={lookerRefCallback} />
-                )}
-                {jsonPanel.isOpen && (
-                  <JSONPanel
-                    containerRef={jsonPanel.containerRef}
-                    onClose={() => jsonPanel.close()}
-                    onCopy={() => jsonPanel.copy()}
-                    json={jsonPanel.json}
-                  />
-                )}
-                {helpPanel.isOpen && (
-                  <HelpPanel
-                    containerRef={helpPanel.containerRef}
-                    onClose={() => helpPanel.close()}
-                    items={helpPanel.items}
-                  />
-                )}
-              </Suspense>
-            </ErrorBoundary>
-          </ContentColumn>
-          <Sidebar render={renderEntry} modal={true} />
-          <OperatorPromptArea area={OPERATOR_PROMPT_AREAS.DRAWER_RIGHT} />
-        </Container>
-      </ModalWrapper>
-    </Fragment>,
+        <OperatorPromptArea area={OPERATOR_PROMPT_AREAS.DRAWER_LEFT} />
+        <TooltipInfo />
+        <SpacesContainer>
+          <Space node={spaces.root} id={MODAL_SPACES_ID} type="modal" />
+        </SpacesContainer>
+        <Sidebar render={renderEntry} modal={true} />
+        <OperatorPromptArea area={OPERATOR_PROMPT_AREAS.DRAWER_RIGHT} />
+      </ModalContainer>
+    </ModalWrapper>,
     document.getElementById("modal") as HTMLDivElement
   );
 };
 
-export default React.memo(SampleModal);
+export default React.memo(Modal);
